+++ src/main/java/io/github/coachluck/backpacksplus/BackPacksPlus.java
@@ -1,6 +1,6 @@
 /*
  *     File: BackPacksPlus.java
- *     Last Modified: 1/13/21, 4:37 PM
+ *     Last Modified: 1/13/21, 10:49 PM
  *     Project: BackPacksPlus
  *     Copyright (C) 2020 CoachL_ck
  *
@@ -20,9 +20,9 @@
 
 package io.github.coachluck.backpacksplus;
 
+import io.github.coachluck.backpacksplus.api.Timer;
 import io.github.coachluck.backpacksplus.utils.BackPack;
 import io.github.coachluck.backpacksplus.utils.InventoryWatcher;
-import io.github.coachluck.backpacksplus.utils.Timer;
 import io.github.coachluck.backpacksplus.utils.backend.Backend;
 import io.github.coachluck.backpacksplus.utils.backend.ChatUtil;
 import io.github.coachluck.backpacksplus.utils.lang.MessageService;
