+++ src/main/java/io/github/coachluck/backpacksplus/api/DisplayItemHelper.java
@@ -1,6 +1,6 @@
 /*
  *     File: DisplayItemHelper.java
- *     Last Modified: 9/4/20, 4:55 PM
+ *     Last Modified: 1/13/21, 10:48 PM
  *     Project: BackPacksPlus
  *     Copyright (C) 2020 CoachL_ck
  *
@@ -18,8 +18,9 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package io.github.coachluck.backpacksplus.utils;
+package io.github.coachluck.backpacksplus.api;
 
+import io.github.coachluck.backpacksplus.utils.BackPack;
 import io.github.coachluck.backpacksplus.utils.backend.ChatUtil;
 import io.github.coachluck.backpacksplus.utils.multiversion.ReflectionUtil;
 import net.md_5.bungee.api.chat.BaseComponent;
